Implement subtle::ConstantTimeEq on x25519 secret types
This is already implemented on `curve25519::Scalar`, we just have
to forward it around the newtype wrapper.

There is no way for the user to do this otherwise for some of these,
since the newtype wrapper doesn't allow access to the bytes of the
secret in case of ephemeral secret.